Move emails to a separate field validator and the Contact Object into a generic OAS validator
