+++ amati/fields/email.py
@@ -1,21 +1,19 @@
 """
-Validates the Open API Specification contact object - §4.8.3
+Validates an email according to the RFC5322 ABNF grammar - §3:
 """
 
 from typing import Annotated, Optional
 
 from abnf.grammars import rfc5322
-from pydantic import AfterValidator, AnyUrl
+from pydantic import AfterValidator
 
-from amati.validators import title
 from amati.validators.reference_object import Reference, ReferenceModel
-from amati.validators.generic import GenericObject
 
 
 reference: Reference = ReferenceModel(
-    title=title,
-    url='https://spec.openapis.org/oas/latest.html#contact-object',
-    section='Contact Object'
+    title='Internet Message Format',
+    url='https://www.rfc-editor.org/rfc/rfc5322#section-3',
+    section='Syntax'
 )
 
 
@@ -37,9 +35,3 @@ Email = Annotated[
     Optional[str],
     AfterValidator(_validate_after_email)
 ]
-
-class ContactObject(GenericObject):
-    name: Optional[str] = None
-    url: Optional[AnyUrl] = None
-    email: Optional[Email] = None
-    _reference: Reference =  reference # type: ignore
