+++ amati/validators/contact_object.py
@@ -1,22 +1,11 @@
-"""
-Validates the Open API Specification contact object - §4.8.3
-"""
 
-from typing import Annotated, Optional
+
 
 from abnf.grammars import rfc5322
-from pydantic import AfterValidator, AnyUrl
-
-from amati.validators import title
-from amati.validators.reference_object import Reference, ReferenceModel
-from amati.validators.generic import GenericObject
 
 
-reference: Reference = ReferenceModel(
-    title=title,
-    url='https://spec.openapis.org/oas/latest.html#contact-object',
-    section='Contact Object'
-)
+
+
 
 
 def _validate_after_email(value: str) -> str:
@@ -38,8 +27,4 @@ Email = Annotated[
     AfterValidator(_validate_after_email)
 ]
 
-class ContactObject(GenericObject):
-    name: Optional[str] = None
-    url: Optional[AnyUrl] = None
-    email: Optional[Email] = None
-    _reference: Reference =  reference # type: ignore
+
